% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FQA_stats.R
\name{cal_half_life}
\alias{cal_half_life}
\title{cal_half_life}
\usage{
cal_half_life(data, cols, max_n = 100)
}
\arguments{
\item{data}{dt}

\item{cols}{colums}

\item{max_n}{max autopcor lag}
}
\description{
cal_half_life
}
