% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FQA_stats.R
\name{cluster_select}
\alias{cluster_select}
\title{cluster_select}
\usage{
cluster_select(
  summaries,
  data,
  target_num,
  sample_num = 500000L,
  abs_cor_mat = TRUE,
  method = "ward.D2"
)
}
\arguments{
\item{summaries}{summaries}

\item{data}{data.table}

\item{target_num}{target_num}

\item{sample_num}{sample_num}

\item{abs_cor_mat}{abs_cor_mat}

\item{method}{method}
}
\description{
cluster_select
}
