% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FQA_stats.R
\name{compare_data}
\alias{compare_data}
\title{compare}
\usage{
compare_data(left_data, right_data, k = c("ukey", "DataDate", "ticktime"))
}
\arguments{
\item{x}{dt}

\item{y}{dt}

\item{key}{by vec}
}
\description{
compare
}
