% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FQA_stats.R
\name{cor_mat}
\alias{cor_mat}
\title{cor_mat}
\usage{
cor_mat(data, cols, sample_num = 1000000L)
}
\arguments{
\item{data}{dt}

\item{cols}{cols}

\item{sample_num}{sample_num}
}
\description{
Correlation Matrix
}
