% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FQA_stats.R
\name{elimination_select2}
\alias{elimination_select2}
\title{elimination_select2}
\usage{
elimination_select2(pair_cor, all_cols, threshold = 0.9)
}
\arguments{
\item{pair_cor}{cor}

\item{all_cols}{cols}

\item{threshold}{threshold}
}
\description{
elimination_select2
}
