% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FQA_stats.R
\name{elimination_select3}
\alias{elimination_select3}
\title{elimination_select3}
\usage{
elimination_select3(data, all_cols, threshold = 0.9, threads = 1L)
}
\arguments{
\item{all_cols}{cols}

\item{threshold}{threshold}

\item{pair_cor}{cor}
}
\description{
elimination_select3
}
