% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FQA_stats.R
\name{gen_test_data}
\alias{gen_test_data}
\title{generate test data}
\usage{
gen_test_data(ins_num = 2L, num = 1000L)
}
\arguments{
\item{ins_num}{ukey number}

\item{num}{how many data pointers}
}
\description{
generate test data
}
