% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FQA_stats.R
\name{group_mean_pcor}
\alias{group_mean_pcor}
\title{mean pcor by key}
\usage{
group_mean_pcor(dt, f1, f2, by_key)
}
\arguments{
\item{f1}{column1}

\item{f2}{column2}

\item{by_key}{by_key}

\item{data}{dt}
}
\description{
mean pcor by key
}
