% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{select_xx_xy}
\alias{select_xx_xy}
\title{select_xx_xy}
\usage{
select_xx_xy(xx_dt, pcor_sorted_x, xx_threshold = 0.8)
}
\arguments{
\item{xx_dt}{generate by ft[, FM::pcor(.SD), .SDcols= cols]}

\item{pcor_sorted_x, }{x vec sorted by pcor, could has dupe}

\item{xx_threshold, }{pcor threshold}
}
\value{
list
}
\description{
select_xx_xy
}
