% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FQA_stats.R
\name{stats.data.frame}
\alias{stats.data.frame}
\title{stats method for data.frame class}
\usage{
\method{stats}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{Other arguments passed to or from other methods}
}
\description{
stats method for data.frame class
}
