% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FQA_stats.R
\name{stats.numeric}
\alias{stats.numeric}
\title{stats method for numeric class}
\usage{
\method{stats}{numeric}(
  x,
  w = NULL,
  q = c(0, 0.005, 0.01, 0.05, 0.1, 0.25, 0.5, 0.75, 0.9, 0.95, 0.99, 0.995, 1),
  ...
)
}
\arguments{
\item{x}{numeric vector}

\item{w}{weight vector}

\item{q}{quantile vector}

\item{...}{Other arguments passed to or from other methods}
}
\description{
stats method for numeric class
}
